+++ regex.gemspec
@@ -1,11 +1,11 @@
 Gem::Specification.new do |s|
   s.name        = 'easy-regex'
-  s.version     = '0.0.1'
+  s.version     = '1.0.0'
   s.summary     = "Easy regex gem"
   s.description = "Gem to obtain and manipulate the most common regex in a easy way"
   s.authors     = ["Gregorio Vazquez"]
   s.email       = 'gregoriovazya@gmail.com'
-  s.files       = ["lib/easy-regex.rb"]
+  s.files       = ["lib/regex.rb"]
   s.homepage    = ''
   s.license     = 'MIT'
 end
